Moving apollo to root
