+++ src/apollo.ts
@@ -1,5 +1,5 @@
 import { ApolloClient, createHttpLink, InMemoryCache } from '@apollo/client';
-import { storage } from '../../utils/storage';
+import { storage } from '../utils/storage';
 import { setContext } from '@apollo/client/link/context';
 
 const BASE_LINK = 'https://template-onboarding-node-sjz6wnaoia-uc.a.run.app/graphql';
